refactor: move ModuleDir and TfDir to terraform package
